[PORT FROM R3]Battery: Evaluate charge termination current correctly
BZ: 59300
R3-Stable:http://android.intel.com:8080/#/c/73211
Bug 46163 changed the logic of evaluating charge termination current from
FG-table. term_curr field used to store this value is a u16-int.
For LEX, upon using TERMINATION_CUR_CONV_FACTOR, this field overruns 16-bit value.
Hence correcting this to accomodate higher values.
Reviewed-on: http://android.intel.com:8080/77327
Tested-by: cactus <cactus@intel.com>

+++ drivers/power/intel_mdf_battery.c
@@ -2708,7 +2708,7 @@ static void sfi_table_invalid_batt(struct msic_batt_sfi_prop *sfi_table)
 * mfld_umip_read_termination_current - reads the termination current data from umip using IPC.
 * @term_curr : termination current read from umip.
 */
-static void  mfld_umip_read_termination_current(u16 *term_curr)
+static void  mfld_umip_read_termination_current(u32 *term_curr)
 {
 	int mip_offset, ret;
 	/* Read 2bytes of termination current data from the umip */
